# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CounterfactualList": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CounterfactualResult"
            },
            "title": "Results",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "results"
        ],
        "title": "CounterfactualList",
        "type": "object"
      },
      "CounterfactualRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "intervention": {
            "$ref": "#/components/schemas/InterventionKind"
          },
          "parent_role": {
            "default": "parent_a",
            "enum": [
              "parent_a",
              "parent_b"
            ],
            "title": "Parent Role",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "intervention"
        ],
        "title": "CounterfactualRequest",
        "type": "object"
      },
      "CounterfactualResult": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "changed_loci": {
            "items": {
              "type": "string"
            },
            "title": "Changed Loci",
            "type": "array"
          },
          "computational_only": {
            "default": true,
            "title": "Computational Only",
            "type": "boolean"
          },
          "counterfactual": {
            "$ref": "#/components/schemas/PhenotypeResult"
          },
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "intervention": {
            "$ref": "#/components/schemas/InterventionKind"
          },
          "intervention_id": {
            "title": "Intervention Id",
            "type": "string"
          },
          "novelty_resolved": {
            "title": "Novelty Resolved",
            "type": "boolean"
          },
          "original": {
            "$ref": "#/components/schemas/PhenotypeResult"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "intervention_id",
          "run_id",
          "candidate_id",
          "intervention",
          "original",
          "counterfactual",
          "delta",
          "novelty_resolved",
          "changed_loci",
          "explanation"
        ],
        "title": "CounterfactualResult",
        "type": "object"
      },
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "default": "fixture_epistasis_ab_cd_v1",
            "title": "Dataset Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode",
            "default": "synthetic"
          },
          "options": {
            "$ref": "#/components/schemas/RunOptions"
          },
          "phenotype_model_id": {
            "default": "synthetic_height_like",
            "title": "Phenotype Model Id",
            "type": "string"
          },
          "seed": {
            "default": 42,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "DataDisclosure": {
        "additionalProperties": false,
        "properties": {
          "causal_interpretation": {
            "minLength": 1,
            "title": "Causal Interpretation",
            "type": "string"
          },
          "dataset_kind": {
            "minLength": 1,
            "title": "Dataset Kind",
            "type": "string"
          },
          "genotype_source": {
            "minLength": 1,
            "title": "Genotype Source",
            "type": "string"
          },
          "phenotype_source": {
            "minLength": 1,
            "title": "Phenotype Source",
            "type": "string"
          }
        },
        "required": [
          "dataset_kind",
          "genotype_source",
          "phenotype_source",
          "causal_interpretation"
        ],
        "title": "DataDisclosure",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InterventionKind": {
        "enum": [
          "revert_variant",
          "swap_segment",
          "break_interaction"
        ],
        "title": "InterventionKind",
        "type": "string"
      },
...
      "PhenotypeResult": {
        "additionalProperties": false,
        "properties": {
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/PhenotypeContribution"
            },
            "title": "Contributions",
            "type": "array"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "subject_id",
          "value",
          "contributions",
          "model_id",
          "model_version",
          "formula"
        ],
        "title": "PhenotypeResult",
        "type": "object"
      },
...
      "RunCreateResponse": {
        "additionalProperties": false,
        "properties": {
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "snapshot_url": {
            "title": "Snapshot Url",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "events_url",
          "snapshot_url"
        ],
        "title": "RunCreateResponse",
        "type": "object"
      },
...
      "RunMode": {
        "enum": [
          "synthetic",
          "real_trio_synthetic_phenotype"
        ],
        "title": "RunMode",
        "type": "string"
      },
      "RunOptions": {
        "additionalProperties": false,
        "properties": {
          "candidate_interactions": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Candidate Interactions",
            "type": "integer"
          },
          "family_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "include_noise": {
            "default": false,
            "title": "Include Noise",
            "type": "boolean"
          },
          "locus_count": {
            "default": 50,
            "maximum": 1000.0,
            "minimum": 20.0,
            "title": "Locus Count",
            "type": "integer"
          },
          "max_candidates": {
            "default": 50,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Max Candidates",
            "type": "integer"
          },
          "region_chromosome": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Chromosome"
          },
          "region_end": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region End"
          },
          "region_start": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Start"
          }
        },
        "title": "RunOptions",
        "type": "object"
      },
      "RunSnapshot": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/TraceCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "counterfactuals": {
            "items": {
              "$ref": "#/components/schemas/CounterfactualResult"
            },
            "title": "Counterfactuals",
            "type": "array"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "disclosure": {
            "$ref": "#/components/schemas/DataDisclosure"
          },
          "evidence_graph": {
            "$ref": "#/components/schemas/EvidenceGraph"
          },
          "genome": {
            "$ref": "#/components/schemas/GenomeSnapshot"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "phenotype": {
            "$ref": "#/components/schemas/PhenotypeBundle"
          },
          "reproducibility": {
            "$ref": "#/components/schemas/ProvenanceMetadata"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_artifacts": {
            "items": {
              "$ref": "#/components/schemas/SourceArtifactReference"
            },
            "title": "Source Artifacts",
            "type": "array"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "status",
          "mode",
          "seed",
          "stage",
          "dataset_id",
          "genome",
          "phenotype",
          "candidates",
          "evidence_graph",
          "reproducibility",
          "disclosure"
        ],
        "title": "RunSnapshot",
        "type": "object"
      },
      "RunStage": {
        "enum": [
          "setup",
          "parent_loading",
          "meiosis",
          "fertilization",
          "phenotype",
          "novelty_detection",
          "candidate_mining",
          "counterfactual",
          "evidence_graph",
          "complete"
        ],
        "title": "RunStage",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "created",
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "mode",
          "stage",
          "progress",
          "seed",
          "dataset_id",
          "model_version"
        ],
        "title": "RunStatusResponse",
        "type": "object"
      },
...
      "TimelineResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "next_sequence": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Sequence"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "events"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Explainable inheritance simulation with deterministic synthetic phenotype models. Results are computational evidence and are not clinical predictions.",
    "title": "Genetic Universe API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs": {
      "post": {
        "operationId": "create_run_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/counterfactuals": {
      "get": {
        "operationId": "list_counterfactuals_v1_runs__run_id__counterfactuals_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterfactualList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Counterfactuals",
        "tags": [
          "counterfactuals"
        ]
      },
      "post": {
        "operationId": "execute_counterfactual_v1_runs__run_id__counterfactuals_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterfactualRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterfactualResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Execute Counterfactual",
        "tags": [
          "counterfactuals"
        ]
      }
    },
    "/v1/runs/{run_id}/snapshot": {
      "get": {
        "operationId": "snapshot_v1_runs__run_id__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Snapshot",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/start": {
      "post": {
        "operationId": "start_run_v1_runs__run_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/timeline": {
      "get": {
        "operationId": "timeline_v1_runs__run_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Timeline",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}